test: split operator routes coverage
